refactor(organism): rename genome.yaml -> organs_registry.yaml (Step 1/3 IG-3)
WIP step 1: file rename + Python module rename + backward-compat shim.

Changes:
- git mv apps/organism/organism/genome.yaml -> organs_registry.yaml
- git mv apps/organism/organism/tools/validate_genome.py -> validate_organs_registry.py
- git mv apps/organism/tests/tools/test_validate_genome.py -> test_validate_organs_registry.py
- New file: apps/organism/organism/genome.yaml symlink -> organs_registry.yaml
  (filesystem alias, removal scheduled 2026-06-08)
- New file: apps/organism/organism/tools/validate_genome.py shim re-exporting
  validate_organs_registry with DeprecationWarning + CLI stderr note
- Updated .pre-commit-config.yaml hook id validate-genome -> validate-organs-registry
- Updated .secrets.baseline path key
- Updated apps/cell/cell/sensors/genome_aggregator_sensor.py default path
- Updated apps/cell/cell/utils/organ_emitter.py module docstring
- Updated apps/cell/tests/test_genome_aggregator_sensor.py docstring + comment
- Updated apps/organism/organism/control_panel.py heartbeat comment
- Updated apps/organism/tests/test_genome_no_active_active.py path + msg
- Updated apps/organism/README.md with new commands + deprecation note
- Updated organs_registry.yaml header comment + checksum
  (checksum was already stale on origin/main from PR #496 W1.5;
  recomputed in-place to avoid yaml.safe_dump comment loss)

Two new tests for the backward-compat surface:
- test_legacy_validate_genome_alias_warns_but_works
- test_legacy_genome_yaml_symlink_resolves_to_organs_registry

cell-core/Genome Python class UNCHANGED (biological metaphor preserved).
SYMBIOSIS.md "Innervation Genoma" SYSTEM CONCEPT preserved.

Tests: 252 passed in apps/organism/tests/ (excluding test_control_panel.py
which needs fastapi locally; runs in CI). Live registry validates clean.
