+++ .secrets.baseline
@@ -5676,10 +5676,10 @@
         "is_secret": false
       }
     ],
-    "apps/organism/organism/genome.yaml": [
+    "apps/organism/organism/organs_registry.yaml": [
       {
         "type": "Hex High Entropy String",
-        "filename": "apps/organism/organism/genome.yaml",
+        "filename": "apps/organism/organism/organs_registry.yaml",
         "hashed_secret": "0844ade45019d71d2b70affc032ef3a63da49750",
         "is_verified": false,
         "line_number": 31,
